refactor(publisher): extract typed errors into errors.ts
Move the five publisher Error classes (PublisherWalletRequiredError,
StaleWriteError, ReservedNamespaceError, AssertionNotPersistedError,
MultiRootPublishNotAtomicError) and the CASCondition payload out of the
4.5k-line dkg-publisher.ts into a self-contained errors.ts.

dkg-publisher.ts re-exports all of them, so `./dkg-publisher.js`, the package
index (which re-exports from this module), every external/deep import, and all
`instanceof` checks are unchanged. Pure mechanical move — no behavior change.
